% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.RtinyccCall}
\alias{.RtinyccCall}
\title{we are using .Call directly, this is to make R CMD check happy}
\usage{
.RtinyccCall()
}
\value{
The result returned by the invoked native routine. This alias is
package-internal and exists to satisfy \verb{R CMD check} native routine
registration requirements.
}
\description{
we are using .Call directly, this is to make R CMD check happy
}
\keyword{internal}
