% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{generate_trampoline}
\alias{generate_trampoline}
\title{Generate trampoline code for a callback argument}
\usage{
generate_trampoline(trampoline_name, sig)
}
\arguments{
\item{trampoline_name}{Name of the trampoline function}

\item{sig}{Parsed signature}
}
\value{
C code string for the trampoline
}
\description{
Generate trampoline code for a callback argument
}
