% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{is_callback_type}
\alias{is_callback_type}
\title{Check if a type represents a callback}
\usage{
is_callback_type(type)
}
\arguments{
\item{type}{Type string to check}
}
\value{
Logical
}
\description{
Check if a type represents a callback
}
