% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{parse_callback_type}
\alias{parse_callback_type}
\title{Parse callback type specification}
\usage{
parse_callback_type(type)
}
\arguments{
\item{type}{Type string like "callback:double(int,int)"}
}
\value{
Parsed signature list or NULL
}
\description{
Parse callback type specification
}
