% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{print.tcc_compiled}
\alias{print.tcc_compiled}
\title{Print tcc_compiled object}
\usage{
\method{print}{tcc_compiled}(x, ...)
}
\arguments{
\item{x}{A tcc_compiled object}

\item{...}{Ignored}
}
\value{
The input \code{tcc_compiled} object, invisibly. Called for its side
effect of printing the compilation output mode and the status of
compiled callable symbols.
}
\description{
Print tcc_compiled object
}
