% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{print.tcc_cstring}
\alias{print.tcc_cstring}
\title{Print a \code{tcc_cstring} object}
\usage{
\method{print}{tcc_cstring}(x, ...)
}
\arguments{
\item{x}{A \code{tcc_cstring} object.}

\item{...}{Ignored.}
}
\value{
The input \code{tcc_cstring} object, invisibly. Called for its side
effect of printing the current string value.
}
\description{
Print a \code{tcc_cstring} object
}
