% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_add_file}
\alias{tcc_add_file}
\title{Add a source file to a libtcc state}
\usage{
tcc_add_file(state, path)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{path}{Path to a C source file.}
}
\value{
Integer status code (0 = success).
}
\description{
Add a source file to a libtcc state
}
