% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_add_include_path}
\alias{tcc_add_include_path}
\title{Add an include path to a libtcc state}
\usage{
tcc_add_include_path(state, path)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{path}{Path to include directory.}
}
\value{
Integer status code (0 = success).
}
\description{
Add an include path to a libtcc state
}
