% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_add_library}
\alias{tcc_add_library}
\title{Add a library to a libtcc state}
\usage{
tcc_add_library(state, library)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{library}{Library name (e.g., "m" for libm, "R" for libR).}
}
\value{
Integer status code (0 = success).
}
\description{
Add a library to a libtcc state
}
