% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_add_library_path}
\alias{tcc_add_library_path}
\title{Add a library path to a libtcc state}
\usage{
tcc_add_library_path(state, path)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{path}{Path to library directory.}
}
\value{
Integer status code (0 = success).
}
\description{
Add a library path to a libtcc state
}
