% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_add_symbol}
\alias{tcc_add_symbol}
\title{Add a symbol to a libtcc state}
\usage{
tcc_add_symbol(state, name, addr)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{name}{Symbol name.}

\item{addr}{External pointer address or symbol value.}
}
\value{
Integer status code (0 = success).
}
\description{
Add a symbol to a libtcc state
}
