% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_call_symbol_int}
\alias{tcc_call_symbol_int}
\title{Call a zero-argument symbol returning int from a libtcc state}
\usage{
tcc_call_symbol_int(state, name)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{name}{Symbol name to call.}
}
\value{
Integer return value from the symbol.
}
\description{
Call a zero-argument symbol returning int from a libtcc state
}
