% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback_async_schedule}
\alias{tcc_callback_async_schedule}
\title{Schedule a callback to run on the main thread}
\usage{
tcc_callback_async_schedule(callback, args = list())
}
\arguments{
\item{callback}{A tcc_callback object}

\item{args}{List of arguments to pass to the callback}
}
\value{
NULL (invisible)
}
\description{
Enqueue a callback for main-thread execution. Arguments must be basic
scalars or external pointers.
}
