% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback_close}
\alias{tcc_callback_close}
\title{Close/unregister a callback}
\usage{
tcc_callback_close(callback)
}
\arguments{
\item{callback}{A tcc_callback object returned by tcc_callback()}
}
\value{
NULL (invisible)
}
\description{
Invalidates a callback immediately, releases the preserved R function
reference, and cleans up callback resources as early as possible.
This is recommended for deterministic lifetime management, but callbacks
are also eventually released by finalizers if you simply drop all references.
}
