% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback_valid}
\alias{tcc_callback_valid}
\title{Check if callback is still valid}
\usage{
tcc_callback_valid(callback)
}
\arguments{
\item{callback}{A tcc_callback object}
}
\value{
Logical indicating if callback can be invoked
}
\description{
Check if callback is still valid
}
