% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_compile}
\alias{tcc_compile}
\title{Compile FFI bindings}
\usage{
tcc_compile(ffi, verbose = FALSE)
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{verbose}{Print compilation info}
}
\value{
A tcc_compiled object with callable functions
}
\description{
Compile the defined symbols into callable functions.
This generates C wrapper code and compiles it with TinyCC.
}
