% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_compile_string}
\alias{tcc_compile_string}
\title{Compile C code from a character string}
\usage{
tcc_compile_string(state, code)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{code}{C source code string.}
}
\value{
Integer status code (0 = success).
}
\description{
Compile C code from a character string
}
