% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_field_addr}
\alias{tcc_field_addr}
\title{Generate field address getter helpers}
\usage{
tcc_field_addr(ffi, struct_name, fields)
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{struct_name}{Struct name}

\item{fields}{Character vector of field names}
}
\value{
Updated tcc_ffi object
}
\description{
Creates functions that return pointers to specific struct fields.
Useful for passing field pointers to C functions or for container_of.
}
\examples{
ffi <- tcc_ffi() \%>\%
  tcc_struct(point = list(x = "f64", y = "f64")) \%>\%
  tcc_field_addr("point", c("x", "y"))  # point_x_addr(), point_y_addr()
}
