% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_header}
\alias{tcc_header}
\title{Add C headers}
\usage{
tcc_header(ffi, header)
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{header}{Header string or include directive}
}
\value{
Updated tcc_ffi object (for chaining)
}
\description{
Add C headers
}
