% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_include}
\alias{tcc_include}
\title{Add include path to FFI context}
\usage{
tcc_include(ffi, path)
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{path}{Include directory path}
}
\value{
Updated tcc_ffi object (for chaining)
}
\description{
Add include path to FFI context
}
