% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_library_path}
\alias{tcc_library_path}
\title{Add library path to FFI context}
\usage{
tcc_library_path(ffi, path)
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{path}{Library directory path}
}
\value{
Updated tcc_ffi object (for chaining)
}
\description{
Add library path to FFI context
}
