% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_malloc}
\alias{tcc_malloc}
\title{Allocate memory buffer}
\usage{
tcc_malloc(size)
}
\arguments{
\item{size}{Number of bytes to allocate}
}
\value{
An external pointer tagged \code{"rtinycc_owned"} with an R finalizer.
Freed on garbage collection or explicitly via \code{\link[=tcc_free]{tcc_free()}}.
}
\description{
Allocate a memory buffer of specified size, equivalent to C malloc.
Returns an external pointer that can be passed to FFI functions.
}
