% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_prefix}
\alias{tcc_prefix}
\alias{tcc_lib_path}
\alias{tcc_include_path}
\alias{tcc_bin_path}
\alias{tcc_cli}
\title{TinyCC paths}
\usage{
tcc_prefix()

tcc_lib_path()

tcc_include_path()

tcc_bin_path()

tcc_cli()
}
\value{
A character scalar path.
}
\description{
Helpers to locate the bundled tinycc installation after the package is installed.
}
