% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_ptr_is_null}
\alias{tcc_ptr_is_null}
\title{Check whether an external pointer is NULL}
\usage{
tcc_ptr_is_null(ptr)
}
\arguments{
\item{ptr}{External pointer}
}
\value{
Logical scalar
}
\description{
Returns TRUE if the external pointer address is NULL, FALSE otherwise.
}
