% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_ptr_is_owned}
\alias{tcc_ptr_is_owned}
\title{Check whether an external pointer is owned by Rtinycc}
\usage{
tcc_ptr_is_owned(ptr)
}
\arguments{
\item{ptr}{External pointer}
}
\value{
Logical scalar
}
\description{
Check whether an external pointer is owned by Rtinycc
}
