% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_read_bytes}
\alias{tcc_read_bytes}
\title{Read raw bytes from a pointer}
\usage{
tcc_read_bytes(ptr, nbytes)
}
\arguments{
\item{ptr}{External pointer}

\item{nbytes}{Number of bytes to read}
}
\value{
Raw vector
}
\description{
Read a fixed number of bytes from an external pointer into a raw vector.
}
