% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_read_f32}
\alias{tcc_read_f32}
\title{Read 32-bit float}
\usage{
tcc_read_f32(ptr, offset = 0L)
}
\arguments{
\item{ptr}{External pointer}

\item{offset}{Byte offset from \code{ptr} (scalar interface). Ignored when
\code{n} is supplied.}
}
\value{
Numeric scalar (promoted to double).
}
\description{
Read 32-bit float
}
