% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_read_f64}
\alias{tcc_read_f64}
\title{Read 64-bit doubles from a pointer}
\usage{
tcc_read_f64(ptr, n)
}
\arguments{
\item{ptr}{External pointer}

\item{n}{Number of values to read}
}
\value{
Numeric vector
}
\description{
Read 64-bit doubles from a pointer
}
