% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_read_i8}
\alias{tcc_read_i8}
\title{Read signed 8-bit integer}
\usage{
tcc_read_i8(ptr, offset = 0L)
}
\arguments{
\item{ptr}{External pointer}

\item{offset}{Byte offset from \code{ptr} (scalar interface). Ignored when
\code{n} is supplied.}
}
\value{
Integer scalar
}
\description{
Read signed 8-bit integer
}
