% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_run_cli}
\alias{tcc_run_cli}
\title{Run the tinycc CLI}
\usage{
tcc_run_cli(args = character(), tcc_path = check_cli_exists())
}
\arguments{
\item{args}{Character vector of CLI arguments (e.g., \code{c("-c", file, "-o", out)}).}

\item{tcc_path}{Optional path to the \code{tcc} binary; defaults to the bundled CLI.}
}
\value{
Integer status from \code{system2()}.
}
\description{
Run the tinycc CLI
}
