% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_source}
\alias{tcc_source}
\title{Add C source code}
\usage{
tcc_source(ffi, code)
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{code}{C source code string}
}
\value{
Updated tcc_ffi object (for chaining)
}
\description{
Add C source code
}
