% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_helpers.R
\name{tcc_treesitter_defines}
\alias{tcc_treesitter_defines}
\title{Extract macro defines from a header file}
\usage{
tcc_treesitter_defines(
  file,
  use_cpp = TRUE,
  cc = treesitter.c::r_cc(),
  ccflags = NULL
)
}
\arguments{
\item{file}{Path to a header file.}

\item{use_cpp}{Logical; use the C preprocessor if available.}

\item{cc}{Compiler string; passed to \code{system2()} if \code{use_cpp = TRUE}.}

\item{ccflags}{Additional flags for the compiler.}
}
\value{
Character vector of macro names defined in \code{file}.
}
\description{
Extract macro defines from a header file
}
\examples{
\dontrun{
tcc_treesitter_defines("/usr/include/math.h")
}
}
