% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_helpers.R
\name{tcc_treesitter_enum_bindings}
\alias{tcc_treesitter_enum_bindings}
\title{Apply tcc_enum() bindings from a header}
\usage{
tcc_treesitter_enum_bindings(ffi, header, constants = NULL)
}
\arguments{
\item{ffi}{A tcc_ffi object.}

\item{header}{Character scalar containing C declarations.}

\item{constants}{Named list of enum constants.}
}
\value{
Updated tcc_ffi object.
}
\description{
Apply tcc_enum() bindings from a header
}
\examples{
header <- "enum status { OK = 0, ERR = 1 };"
ffi <- tcc_ffi()
ffi <- tcc_treesitter_enum_bindings(ffi, header, constants = list(status = c("OK", "ERR")))
}
