% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_helpers.R
\name{tcc_treesitter_struct_bindings}
\alias{tcc_treesitter_struct_bindings}
\title{Apply tcc_struct() bindings from a header}
\usage{
tcc_treesitter_struct_bindings(ffi, header, ...)
}
\arguments{
\item{ffi}{A tcc_ffi object.}

\item{header}{Character scalar containing C declarations.}

\item{...}{Passed to \code{tcc_treesitter_struct_accessors()}.}
}
\value{
Updated tcc_ffi object.
}
\description{
Apply tcc_struct() bindings from a header
}
\examples{
header <- "struct point { double x; double y; };"
ffi <- tcc_ffi()
ffi <- tcc_treesitter_struct_bindings(ffi, header)
}
