% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_helpers.R
\name{tcc_treesitter_union_members}
\alias{tcc_treesitter_union_members}
\title{Parse union members with treesitter.c}
\usage{
tcc_treesitter_union_members(header, ...)
}
\arguments{
\item{header}{Character scalar containing C declarations.}

\item{...}{Additional arguments passed to \code{treesitter.c::get_union_members_from_root()}.}
}
\value{
A data frame of union members.
}
\description{
Parse union members with treesitter.c
}
\examples{
\dontrun{
header <- "union data { int i; double d; };"
tcc_treesitter_union_members(header)
}
}
