% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter_helpers.R
\name{tcc_treesitter_unions}
\alias{tcc_treesitter_unions}
\title{Parse union declarations with treesitter.c}
\usage{
tcc_treesitter_unions(header, ...)
}
\arguments{
\item{header}{Character scalar containing C declarations.}

\item{...}{Additional arguments passed to \code{treesitter.c::get_union_nodes()}.}
}
\value{
A data frame of union nodes.
}
\description{
Parse union declarations with treesitter.c
}
\examples{
\dontrun{
header <- "union data { int i; double d; };"
tcc_treesitter_unions(header)
}
}
