% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_write_bytes}
\alias{tcc_write_bytes}
\title{Write raw bytes to a pointer}
\usage{
tcc_write_bytes(ptr, raw)
}
\arguments{
\item{ptr}{External pointer}

\item{raw}{Raw vector to write}
}
\value{
NULL (invisibly)
}
\description{
Write a raw vector into memory pointed to by an external pointer.
}
