% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_write_f32}
\alias{tcc_write_f32}
\title{Write a 32-bit float}
\usage{
tcc_write_f32(ptr, offset, value)
}
\arguments{
\item{ptr}{External pointer}

\item{offset}{Byte offset}

\item{value}{Integer value to write}
}
\description{
Write a 32-bit float
}
