% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_write_f64}
\alias{tcc_write_f64}
\title{Write a 64-bit double}
\usage{
tcc_write_f64(ptr, offset, value)
}
\arguments{
\item{ptr}{External pointer}

\item{offset}{Byte offset}

\item{value}{Integer value to write}
}
\description{
Write a 64-bit double
}
