% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_write_ptr}
\alias{tcc_write_ptr}
\title{Write a pointer at byte offset}
\usage{
tcc_write_ptr(ptr, offset, value)
}
\arguments{
\item{ptr}{External pointer (destination buffer)}

\item{offset}{Byte offset}

\item{value}{External pointer to write}
}
\value{
\code{NULL} (invisibly).
}
\description{
Write a pointer at byte offset
}
