% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_write_u8}
\alias{tcc_write_u8}
\title{Write an unsigned 8-bit integer}
\usage{
tcc_write_u8(ptr, offset, value)
}
\arguments{
\item{ptr}{External pointer}

\item{offset}{Byte offset}

\item{value}{Integer value to write}
}
\description{
Write an unsigned 8-bit integer
}
