% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.RtinyccCall}
\alias{.RtinyccCall}
\title{we are using .Call directly, this is to make R CMD check happy}
\usage{
.RtinyccCall()
}
\description{
we are using .Call directly, this is to make R CMD check happy
}
