% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{get_external_ptr_addr}
\alias{get_external_ptr_addr}
\title{Get the address of an external pointer}
\usage{
get_external_ptr_addr(ptr)
}
\arguments{
\item{ptr}{External pointer}
}
\value{
Address as numeric
}
\description{
Get the address of an external pointer
}
