% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{print.tcc_callback}
\alias{print.tcc_callback}
\title{Print tcc_callback object}
\usage{
\method{print}{tcc_callback}(x, ...)
}
\arguments{
\item{x}{A tcc_callback object}

\item{...}{Ignored}
}
\description{
Print tcc_callback object
}
