% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{print.tcc_compiled}
\alias{print.tcc_compiled}
\title{Print tcc_compiled object}
\usage{
\method{print}{tcc_compiled}(x, ...)
}
\arguments{
\item{x}{A tcc_compiled object}

\item{...}{Ignored}
}
\description{
Print tcc_compiled object
}
