% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback}
\alias{tcc_callback}
\title{Register an R function as a callback}
\usage{
tcc_callback(fun, signature, threadsafe = FALSE)
}
\arguments{
\item{fun}{An R function to be called from C}

\item{signature}{C function signature string (e.g., "double (*)(int, double)")}

\item{threadsafe}{Whether to enable thread-safe invocation (experimental)}
}
\value{
A tcc_callback object (externalptr wrapper)
}
\description{
Wraps an R function so it can be passed as a C function pointer
to compiled code. The callback will be invoked via a trampoline
that marshals arguments between C and R.
}
