% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback_async_drain}
\alias{tcc_callback_async_drain}
\title{Drain the async callback queue}
\usage{
tcc_callback_async_drain()
}
\value{
NULL (invisible)
}
\description{
This is mainly useful for tests; normally callbacks are executed by the
event loop once scheduled.
}
