% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback_async_enable}
\alias{tcc_callback_async_enable}
\title{Enable async callback dispatcher (main-thread queue)}
\usage{
tcc_callback_async_enable()
}
\value{
NULL (invisible)
}
\description{
Initializes an event-loop handler so callbacks can be scheduled from
non-R threads and executed on the main R thread. This is currently
supported on Unix-like systems only.
}
