% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback_close}
\alias{tcc_callback_close}
\title{Close/unregister a callback}
\usage{
tcc_callback_close(callback)
}
\arguments{
\item{callback}{A tcc_callback object returned by tcc_callback()}
}
\value{
NULL (invisible)
}
\description{
Releases the preserved R function reference and cleans up resources.
Must be called when the callback is no longer needed.
}
