% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{tcc_callback_ptr}
\alias{tcc_callback_ptr}
\title{Get the C-compatible function pointer}
\usage{
tcc_callback_ptr(callback)
}
\arguments{
\item{callback}{A tcc_callback object}
}
\value{
An external pointer (address of the callback token)
}
\description{
Returns an external pointer that can be passed to compiled C code
as a function pointer. This is the opaque pointer used by trampolines.
}
