% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_free}
\alias{tcc_free}
\title{Free allocated memory}
\usage{
tcc_free(ptr)
}
\arguments{
\item{ptr}{External pointer to free}
}
\value{
NULL (invisibly)
}
\description{
Free memory allocated with tcc_malloc, equivalent to C free.
}
