% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_get_symbol}
\alias{tcc_get_symbol}
\title{Get a symbol pointer from a libtcc state}
\usage{
tcc_get_symbol(state, name)
}
\arguments{
\item{state}{A \code{tcc_state}.}

\item{name}{Symbol name to look up.}
}
\value{
External pointer of class \code{tcc_symbol}.
}
\description{
Get a symbol pointer from a libtcc state
}
