% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_library}
\alias{tcc_library}
\title{Add library to link against}
\usage{
tcc_library(ffi, library)
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{library}{Library name (e.g., "m", "sqlite3")}
}
\value{
Updated tcc_ffi object (for chaining)
}
\description{
Add library to link against
}
