% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_list_symbols}
\alias{tcc_list_symbols}
\title{List all symbols in a libtcc state}
\usage{
tcc_list_symbols(state)
}
\arguments{
\item{state}{A \code{tcc_state}.}
}
\value{
Character vector of symbol names.
}
\description{
List all symbols in a libtcc state
}
