% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi.R
\name{tcc_output}
\alias{tcc_output}
\title{Set output type for FFI compilation}
\usage{
tcc_output(ffi, output = c("memory", "dll", "exe"))
}
\arguments{
\item{ffi}{A tcc_ffi object}

\item{output}{One of "memory", "dll", "exe"}
}
\value{
Updated tcc_ffi object (for chaining)
}
\description{
Set output type for FFI compilation
}
