% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_ptr_utils.R
\name{tcc_read_i32}
\alias{tcc_read_i32}
\title{Read signed 32-bit integers from a pointer}
\usage{
tcc_read_i32(ptr, n)
}
\arguments{
\item{ptr}{External pointer}

\item{n}{Number of values to read}
}
\value{
Integer vector
}
\description{
Read signed 32-bit integers from a pointer
}
