% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_relocate}
\alias{tcc_relocate}
\title{Relocate compiled code}
\usage{
tcc_relocate(state)
}
\arguments{
\item{state}{A \code{tcc_state}.}
}
\value{
Integer status code (0 = success).
}
\description{
Relocate compiled code
}
