% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycc.R
\name{tcc_state}
\alias{tcc_state}
\title{Create a libtcc state}
\usage{
tcc_state(
  output = c("memory", "obj", "dll", "exe", "preprocess"),
  include_path = tcc_include_paths(),
  lib_path = tcc_lib_paths()
)
}
\arguments{
\item{output}{Output type: one of "memory", "obj", "dll", "exe", "preprocess".}

\item{include_path}{Path(s) to headers; defaults to the bundled include dirs.}

\item{lib_path}{Path(s) to libraries; defaults to the bundled lib dirs (lib and lib/tcc).}
}
\value{
An external pointer of class \code{tcc_state}.
}
\description{
Initialize a libtcc compilation state, optionally pointing at the bundled include/lib paths.
}
